changed chat to room and /chats to /rooms
