+++ src/App.js
@@ -23,7 +23,7 @@ app.use('/api', routes.login)
 
 /* Protected routes below */
 app.use(auth)
-app.use('/api', routes.chats)
+app.use('/api', routes.rooms)
 
 /* */
 app.use((req, res) => {
