--- src/models/Chat.js
@@ -1,10 +1,0 @@
-import mongoose from "mongoose"
-
-export const Chat = mongoose.model("Chat", {
-  name: {
-    type: String,
-    required: true,
-    unique: true,
-    created: new Date(),
-  },
-})
